Crate vortex_dtype

Re-exports

pub use half;

Modules

field

flatbuffers

proto

Macros

match_each_float_ptype

match_each_integer_ptype

match_each_native_ptype

match_each_unsigned_integer_ptype

Structs

ExtDType

ExtID

ExtMetadata

FieldInfo

StructDType

Enums

DType

Array logical types.

Nullability

Whether an item can contain a null value or not

PType

Physical type enum, represents the in-memory physical layout but might represent a different logical type.

Traits

NativePType

ToBytes

TryFromBytes

Type Aliases

FieldName

FieldNames

Metadata