Enum vortex_dtype::DType

pub enum DType {
    Null,
    Bool(Nullability),
    Primitive(PType, Nullability),
    Utf8(Nullability),
    Binary(Nullability),
    Struct(StructDType, Nullability),
    List(Arc<DType>, Nullability),
    Extension(ExtDType, Nullability),
}

Array logical types.

Vortex arrays preserve a single logical type, while the encodings allow for multiple physical types to encode that type.

Variants

Null

Bool(Nullability)

Primitive(PType, Nullability)

Utf8(Nullability)

Binary(Nullability)

Struct(StructDType, Nullability)

List(Arc<DType>, Nullability)

Extension(ExtDType, Nullability)

Implementations

impl DType

pub const BYTES: Self = _

pub const IDX: Self = _

The default DType for indices

pub fn nullability(&self) -> Nullability

pub fn is_nullable(&self) -> bool

pub fn as_nonnullable(&self) -> Self

pub fn as_nullable(&self) -> Self

pub fn with_nullability(&self, nullability: Nullability) -> Self

pub fn eq_ignore_nullability(&self, other: &Self) -> bool

pub fn is_struct(&self) -> bool

pub fn is_unsigned_int(&self) -> bool

pub fn is_signed_int(&self) -> bool

pub fn is_int(&self) -> bool

pub fn is_float(&self) -> bool

pub fn is_boolean(&self) -> bool

pub fn as_struct(&self) -> Option<&StructDType>