Module vortex_dtype::flatbuffers

Structs

Binary

BinaryArgs

BinaryBuilder

Bool

BoolArgs

BoolBuilder

DType

DTypeArgs

DTypeBuilder

Decimal

DecimalArgs

DecimalBuilder

Extension

ExtensionArgs

ExtensionBuilder

List

ListArgs

ListBuilder

Null

NullArgs

NullBuilder

PType

Primitive

PrimitiveArgs

PrimitiveBuilder

Struct_

Struct_Args

Struct_Builder

Type

TypeUnionTableOffset

Utf8

Utf8Args

Utf8Builder

Enums

BinaryOffset

BoolOffset

DTypeOffset

DecimalOffset

ExtensionOffset

ListOffset

NullOffset

PrimitiveOffset

Struct_Offset

Utf8Offset

Constants

ENUM_MAX_PTYPEDeprecated

ENUM_MAX_TYPEDeprecated

ENUM_MIN_PTYPEDeprecated

ENUM_MIN_TYPEDeprecated

ENUM_VALUES_PTYPEDeprecated

ENUM_VALUES_TYPEDeprecated

Functions

deserialize_and_project

Deserialize flatbuffer schema selecting only columns defined by projection

finish_dtype_buffer

finish_size_prefixed_dtype_buffer

resolve_field

Convert name references in projection list into index references.

root_as_dtype

Verifies that a buffer of bytes contains a DType and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_dtype_unchecked.

root_as_dtype_unchecked⚠

Assumes, without verification, that a buffer of bytes contains a DType and returns it.

root_as_dtype_with_opts

Verifies, with the given options, that a buffer of bytes contains a DType and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_dtype_unchecked.

size_prefixed_root_as_dtype

Verifies that a buffer of bytes contains a size prefixed DType and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use size_prefixed_root_as_dtype_unchecked.

size_prefixed_root_as_dtype_unchecked⚠

Assumes, without verification, that a buffer of bytes contains a size prefixed DType and returns it.

size_prefixed_root_as_dtype_with_opts

Verifies, with the given verifier options, that a buffer of bytes contains a size prefixed DType and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_dtype_unchecked.