Function vortex_dtype::flatbuffers::resolve_field

pub fn resolve_field<'a, 'b: 'a>(
    fb: Struct_<'b>,
    field: &'a Field,
) -> VortexResult<usize>

Convert name references in projection list into index references.

This is mostly useful if you want to deduplicate multiple projections against serialized schema.